+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Change Log
 
 - Fixed possible IndentationError problem with sending modules in Flame
 - Can now deal with exceptions that can't be serialized: they're raised as generic PyroError instead with appropriate message
+- added new config item FLAME_ENABLED, to enable/disable the use of Pyro Flame on the server. Default is false (disabled).
 
 
 **Pyro 4.16**